+++ tests/test_correlator_context.py
@@ -3,6 +3,8 @@ from os.path import join as path_join, dirname
 import numpy as np
 import pytest
 
+from pymwalib.common import MWAVersion
+from pymwalib.metafits_context import MetafitsContext
 from pymwalib.correlator_context import CorrelatorContext
 
 
@@ -26,6 +28,19 @@ def mwax_corr_context() -> CorrelatorContext:
     )
 
 
+def test_metafits_context():
+    context = MetafitsContext(prefix_test_data("1297526432_mwax/1297526432.metafits"), None)
+    assert context.mwa_version == MWAVersion.CorrLegacy.value
+
+
+def test_mwax_metafits_context(mwax_corr_context: CorrelatorContext):
+    assert mwax_corr_context.metafits_context.obs_id == 1297526432
+    assert mwax_corr_context.mwa_version.value == MWAVersion.CorrMWAXv2.value
+    assert mwax_corr_context.metafits_context.mwa_version == MWAVersion.CorrMWAXv2.value
+    assert len(mwax_corr_context.metafits_context.receivers) == mwax_corr_context.metafits_context.num_receivers
+    assert len(mwax_corr_context.metafits_context.delays) == mwax_corr_context.metafits_context.num_delays
+
+
 def test_mwax_antennas(mwax_corr_context: CorrelatorContext):
     assert len(mwax_corr_context.metafits_context.antennas) == 2
     assert mwax_corr_context.metafits_context.antennas[0].tile_id == 51
@@ -46,6 +61,12 @@ def test_mwax_rfinputs(mwax_corr_context: CorrelatorContext):
     assert len(mwax_corr_context.metafits_context.rf_inputs) == 4
     assert mwax_corr_context.metafits_context.rf_inputs[0].tile_id == 51
     assert mwax_corr_context.metafits_context.rf_inputs[0].pol == "X"
+    assert len(mwax_corr_context.metafits_context.rf_inputs[0].dipole_delays) == \
+           mwax_corr_context.metafits_context.rf_inputs[0].num_dipole_delays
+    assert len(mwax_corr_context.metafits_context.rf_inputs[0].dipole_gains) == \
+           mwax_corr_context.metafits_context.rf_inputs[0].num_dipole_gains
+    assert len(mwax_corr_context.metafits_context.rf_inputs[0].digital_gains) == \
+           mwax_corr_context.metafits_context.rf_inputs[0].num_digital_gains
     assert mwax_corr_context.metafits_context.rf_inputs[1].tile_id == 51
     assert mwax_corr_context.metafits_context.rf_inputs[1].pol == "Y"
     assert mwax_corr_context.metafits_context.rf_inputs[2].tile_id == 52